refactor: organize and updated rest of queries
