--- src/com/bryan/finance/database/queries/Address.java
@@ -1,4 +1,0 @@
-package com.bryan.finance.database.queries;
-
-public class Address {
-}
