+++ src/com/bryan/finance/enums/Tables.java
@@ -1,5 +1,6 @@
 package com.bryan.finance.enums;
 
 public enum Tables {
-	ADDRESSES, EXPENSES, INCOME, INVESTMENTS, MONTHLY_TOTALS, MONTHLY_EXP_CATGR, MONTHLY_TRANSACTIONS, SAVINGS, SITES, USERS
+	ADDRESSES, BANNED_USERS, EXPENSES, INCOME, INVESTMENTS, MONTHLY_TOTALS,
+	MONTHLY_EXP_CATGR, MONTHLY_TRANSACTIONS, PAY_GRADES, SAVINGS, SITES, USERS
 }
