+++ src/com/bryan/finance/gui/SalaryManagement.java
@@ -21,6 +21,7 @@ import javax.swing.border.Border;
 
 import com.bryan.finance.beans.Salary;
 import com.bryan.finance.database.queries.Queries;
+import com.bryan.finance.database.queries.QueryUtil;
 import com.bryan.finance.gui.util.Title;
 import com.bryan.finance.literals.ApplicationLiterals;
 import com.bryan.finance.literals.Icons;
@@ -51,7 +52,7 @@ public class SalaryManagement implements ActionListener {
 	private JLabel totalPayAndBonus = new JLabel();
 
 	public SalaryManagement() {
-		salaries = Queries.getSalaryData();
+		salaries = QueryUtil.getSalaryData();
 
 		populateJobGrades();
 		JLabel title = new Title("Salary Calculator");
