+++ src/com/bryan/finance/gui/address/AddressTab.java
@@ -27,6 +27,7 @@ import javax.swing.event.TableModelEvent;
 import javax.swing.event.TableModelListener;
 import javax.swing.table.DefaultTableModel;
 
+import com.bryan.finance.database.queries.Addresses;
 import org.apache.log4j.Logger;
 
 import com.bryan.finance.beans.Address;
@@ -166,7 +167,7 @@ public class AddressTab extends JPanel {
 	}
 
 	private JScrollPane getAddressData() {
-		Object[][] records = Queries.getAddresses();
+		Object[][] records = Addresses.getAddresses();
 		Object[] columnNames = { "ID", "Last Name", "First Name(s)", "Address",
 				"City", "State", "ZIP" };
 
@@ -219,7 +220,7 @@ public class AddressTab extends JPanel {
 	}
 
 	private JScrollPane getPartialAddressData() {
-		Object[][] records = Queries.getAddresses();
+		Object[][] records = Addresses.getAddresses();
 		Object[][] partialRecords = getFirstTwoDataColumns(records);
 		Object[] columnNames = { "Last Name", "First Name(s)" };
 
@@ -267,7 +268,7 @@ public class AddressTab extends JPanel {
 						firstNameValue = cell;
 					}
 
-					Address address = Queries.getSpecifiedAddress(
+					Address address = Addresses.getSpecifiedAddress(
 							lastNameValue, firstNameValue);
 					new AddressRecord(address);
 				}
@@ -288,7 +289,7 @@ public class AddressTab extends JPanel {
 
 	private Map<Integer, String> getAttributeMap() {
 
-		Map<Integer, String> map = new HashMap<Integer, String>();
+		Map<Integer, String> map = new HashMap<>();
 		map.put(1, "LAST_NAME");
 		map.put(2, "FIRST_NAMES");
 		map.put(3, "ADDRESS");
