+++ src/com/bryan/finance/gui/util/ButtonEditor.java
@@ -9,6 +9,7 @@ import javax.swing.JOptionPane;
 import javax.swing.JTable;
 
 import com.bryan.finance.database.Connect;
+import com.bryan.finance.database.queries.Accounts;
 import org.apache.log4j.Logger;
 
 import com.bryan.finance.database.queries.Queries;
@@ -101,7 +102,7 @@ public class ButtonEditor extends DefaultCellEditor {
 
 	private void lockUser(String user) {
 		logger.info("LOCKING USER: " + user);
-		Queries.lockUser(user);
+		Accounts.lockUser(user);
 		JOptionPane.showMessageDialog(null, user
 						+ " has been locked", "Locked",
 				JOptionPane.INFORMATION_MESSAGE);
@@ -109,7 +110,7 @@ public class ButtonEditor extends DefaultCellEditor {
 
 	private void unlockUser(String user) {
 		logger.info("UNLOCKING USER: " + user);
-		Queries.unlockUser(user);
+		Accounts.unlockUser(user);
 		JOptionPane.showMessageDialog(null, user
 						+ " has been unlocked", "Unlocked",
 				JOptionPane.INFORMATION_MESSAGE);
